Convert a list of sf objects to a single sf object

A fast alternative to do.call(rbind, <list_of_sf>)

st_rbindlist(
  l,
  ...,
  use_geometry = FALSE,
  use_any_geometry = FALSE,
  geometry_name = NULL
)

Arguments

l
list of objects of the class sf

...
arguments of the function rbindlist except l

use_geometry
TRUE detects and binds the active list-columns with simple feature geometries separately if all of them have the same geometry type, thus enabling matching of geometry columns differing by name as well as by position across the listed sf objects.

use_any_geometry
TRUE does the same as use_geometry = TRUE, but allows different geometry types among the listed sf objects. If use_any_geometry = TRUE then use_geometry is ignored.

geometry_name
a single character string giving a user-defined name to the returned active list-column with simple feature geometries.

Value

a sf object

Details

st_rbindlist() is basically a wrapper for sf::st_as_sf(data.table::rbindlist(<list_of_sf>)), which is a fast alternative to do.call(rbind, <list_of_sf>). It comes with some properties which rbind has, but rbindlist misses:

checks if all sf objects included in the input share the same CRS

returns a sf object with a correct bbox (more information on how to handle this rbindlist issue can be found here).

allows binding sf objects with different geometry types, which can be achieved by setting st_rbindlist's argument use_any_geometry to TRUE.

Remarks about matching geometry columns of input sf objects:

If the geometry columns are identically named, but differ by position within each sf object set use.names = TRUE and fill = TRUE.

If the geometry (and other) columns are differently named, but have same position within every sf object set use.names = FALSE.

If the geometry columns are differently named and at the same time differently positioned set use_geometry = TRUE or use_any_geometry = TRUE. This detects the geometry columns and then matches them separately while the other columns can be still treated collectively by the arguments use.names and fill. Note that matching geometry columns separately requires more processing time.

If the argument geometry_name is unspecified (default NULL) the geometry column name of the returned sf object depends on the following:

If separately matched geometry columns (use_geometry = TRUE or use_any_geometry = TRUE) have a common name, this name persists, else "geometry" is used.

If geometry (and other) columns are differently named and matched by their same position (use.names = FALSE), the geometry column name is inherited from the first sf object listed in the input.

See also

Examples

library(magrittr)
library(dplyr)

#> 
#> Attache Paket: ‘dplyr’

#> Die folgenden Objekte sind maskiert von ‘package:raster’:
#> 
#>     intersect, select, union

#> Die folgenden Objekte sind maskiert von ‘package:stats’:
#> 
#>     filter, lag

#> Die folgenden Objekte sind maskiert von ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

library(sf)

# a set of geometries to work with
nc <- st_read(system.file("gpkg/nc.gpkg", package = "sf"), quiet = TRUE)

# proof of concept with a list of sf-objects having identically named and
# positioned attribute and geometry columns:
list_of_sf <- lapply(1:nrow(nc), function(x) nc[x, ])
nc_new <- st_rbindlist(list_of_sf)
all.equal(nc, nc_new)

#> [1] TRUE

# list of sf-objects having differently named, but identically positioned
# geometry columns (attribute columns are also same-positioned and still
# same-named):
list_of_sf <- lapply(
  1:nrow(nc),
  function(x) st_rename_geometry(nc[x, ], paste0("geom_", x))
  )
# such a list of sf-objects is easy to stack ...
nc_new_2 <- st_rbindlist(list_of_sf, use.name = FALSE)
# ... but by default the name of the returned geometry column ...
attr(nc_new_2, "sf_column")

#> [1] "geom_1"

# ... is inherited from the first listed sf-object.
attr(nc_new_2, "sf_column") == attr(list_of_sf[[1]], "sf_column")

#> [1] TRUE

# user defined name of the stacked geometry column:
nc_new_3 <- st_rbindlist(list_of_sf, use.name = FALSE, geometry_name = "geom")
attr(nc_new_3, "sf_column")

#> [1] "geom"

# from which listed element / sf-object do the stacked features originate?
l <- list(A = nc[1, 1:3], B = NULL, C = nc[2:3, 1:3])
st_rbindlist(l, idcol = "ID")

#> Simple feature collection with 3 features and 4 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -81.74107 ymin: 36.23388 xmax: -80.43531 ymax: 36.58965
#> Geodetic CRS:  NAD27
#>   ID  AREA PERIMETER CNTY_                           geom
#> 1  A 0.114     1.442  1825 MULTIPOLYGON (((-81.47276 3...
#> 2  C 0.061     1.231  1827 MULTIPOLYGON (((-81.23989 3...
#> 3  C 0.143     1.630  1828 MULTIPOLYGON (((-80.45634 3...

# list of sf-objects with identically named, but differently positioned
# geometry and attribute columns (the latter might not be included in
# each listed sf-object) ...
l <- list(nc[1, NULL], nc[2, 1], nc[3, 2:3])
# ... requires use.name = TRUE and fill = TRUE
st_rbindlist(l, use.name = TRUE, fill = TRUE, idcol = TRUE)

#> Simple feature collection with 3 features and 4 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -81.74107 ymin: 36.23388 xmax: -80.43531 ymax: 36.58965
#> Geodetic CRS:  NAD27
#>   .id  AREA PERIMETER CNTY_                           geom
#> 1   1    NA        NA    NA MULTIPOLYGON (((-81.47276 3...
#> 2   2 0.061        NA    NA MULTIPOLYGON (((-81.23989 3...
#> 3   3    NA      1.63  1828 MULTIPOLYGON (((-80.45634 3...

# list of sf-objects with differently named and positioned geometry columns,
# and (if existing) identically named attribute columns ...
l <- lapply(
  seq_along(l),
  function(x) st_rename_geometry(l[[x]], paste0("geom_", x))
  )
sapply(l, attr, "sf_column") # geometry column names of listed sf-objects

#> [1] "geom_1" "geom_2" "geom_3"

# ... requires use_geometry = TRUE for detecting and binding the geometry
# columns (separately from the attribute columns)
st_rbindlist(l, use.name = TRUE, fill = TRUE, idcol = TRUE, use_geometry = TRUE)

#> Simple feature collection with 3 features and 4 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -81.74107 ymin: 36.23388 xmax: -80.43531 ymax: 36.58965
#> Geodetic CRS:  NAD27
#>   .id  AREA PERIMETER CNTY_                       geometry
#> 1   1    NA        NA    NA MULTIPOLYGON (((-81.47276 3...
#> 2   2 0.061        NA    NA MULTIPOLYGON (((-81.23989 3...
#> 3   3    NA      1.63  1828 MULTIPOLYGON (((-80.45634 3...

# list of sf-objects having different geometry types,
#  all columns identically named and positioned ...
l <- list(
  multi_poly = nc[1:2, ],
  no_row     = nc[NULL, ],
  poly       = st_cast(nc[4, ], "POLYGON", warn = FALSE),
  empty_row  = nc[nrow(nc) + 1, ]
  )
sapply(l, st_geometry_type, by_geometry = FALSE)

#>   multi_poly       no_row         poly    empty_row 
#> MULTIPOLYGON     GEOMETRY      POLYGON     GEOMETRY 
#> 18 Levels: GEOMETRY POINT LINESTRING POLYGON MULTIPOINT ... TRIANGLE

# ... requires use_any_geometry = TRUE
st_obj <- st_rbindlist(l, idcol = TRUE, use_any_geometry = TRUE)
st_obj[, ".id"] %>% mutate(geometry_type = st_geometry_type(.))

#> Simple feature collection with 6 features and 2 fields (with 1 geometry empty)
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: -81.74107 ymin: 36.07282 xmax: -75.77316 ymax: 36.58965
#> Geodetic CRS:  NAD27
#>          .id                           geom      geometry_type
#> 1 multi_poly MULTIPOLYGON (((-81.47276 3...       MULTIPOLYGON
#> 2 multi_poly MULTIPOLYGON (((-81.23989 3...       MULTIPOLYGON
#> 3       poly POLYGON ((-76.00897 36.3196...            POLYGON
#> 4       poly POLYGON ((-76.02717 36.5567...            POLYGON
#> 5       poly POLYGON ((-75.90199 36.5562...            POLYGON
#> 6  empty_row       GEOMETRYCOLLECTION EMPTY GEOMETRYCOLLECTION

# list of sf-objects having different geometry types as well as having
# differently named and positioned geometry columns, and (if existing)
# identically named attribute columns ...
l <- lapply(
  seq_along(l),
  function(x) {st_rename_geometry(l[[x]], paste0("geom_", x)) %>% .[,1:x]}
  ) %>%
  setNames(names(l))
# ... requires use.name, fill and use_any_geometry all being set to TRUE
st_rbindlist(l, use.name = TRUE, fill = TRUE, idcol = TRUE, use_any_geometry = TRUE)

#> Simple feature collection with 6 features and 5 fields (with 1 geometry empty)
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: -81.74107 ymin: 36.07282 xmax: -75.77316 ymax: 36.58965
#> Geodetic CRS:  NAD27
#>          .id  AREA PERIMETER CNTY_ CNTY_ID                       geometry
#> 1 multi_poly 0.114        NA    NA      NA MULTIPOLYGON (((-81.47276 3...
#> 2 multi_poly 0.061        NA    NA      NA MULTIPOLYGON (((-81.23989 3...
#> 3       poly 0.070     2.968  1831      NA POLYGON ((-76.00897 36.3196...
#> 4       poly 0.070     2.968  1831      NA POLYGON ((-76.02717 36.5567...
#> 5       poly 0.070     2.968  1831      NA POLYGON ((-75.90199 36.5562...
#> 6  empty_row    NA        NA    NA      NA       GEOMETRYCOLLECTION EMPTY